requirements.txt
apache-airflow==2.5.1
apache-airflow-providers-postgres==5.1.0
apache-airflow-providers-mysql==3.1.0
pandas==1.5.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
mysql-connector-python==8.0.32